pyproject.toml
[tool.poetry]
name = "humidifier"
version = "0.1.0"
description = ""
authors = ["fahreeve@yandex.ru"]
license = "MIT"
package-mode = false

[tool.poetry.dependencies]
python = "^3.13"
python-miio = {git = "https://github.com/rytilahti/python-miio.git"}
aiomqtt = "^2.4.0"

[poetry.group.dev.dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
